Create a new module for subexperiment generation
