+++ circuit_knitting/cutting/__init__.py
@@ -90,8 +90,8 @@ from .cutting_decomposition import (
 from .cutting_evaluation import (
     execute_experiments,
     CuttingExperimentResults,
-    generate_cutting_experiments,
 )
+from .cutting_experiments import generate_cutting_experiments
 from .cutting_reconstruction import reconstruct_expectation_values
 from .wire_cutting_transforms import cut_wires, expand_observables
